Changed all instances of decimal to double.

+++ VideoEditorWPF/MainWindow.xaml.cs
@@ -90,12 +90,12 @@ namespace VideoEditorWPF
             timelineView.ScaleFactor = newScaleFactor;
 		}
 
-        private void timelineView_eventMoved(TimelineEvent timelineEvent, decimal newStartTime)
+        private void timelineView_eventMoved(TimelineEvent timelineEvent, double newStartTime)
         {
             //TODO: Error checking
 
             //Move the event
-            decimal delta = newStartTime - timelineEvent.startTime;
+            double delta = newStartTime - timelineEvent.startTime;
             timelineEvent.startTime += delta;
             timelineEvent.endTime += delta;
         }
